Rename enkf -> ensemble_kalman, update -> filter_update
