+++ cuthbert/__init__.py
@@ -3,7 +3,7 @@ from importlib.metadata import version
 __version__ = version("cuthbert")
 del version
 
-from cuthbert import discrete, enkf, factorial, gaussian, smc
+from cuthbert import discrete, ensemble_kalman, factorial, gaussian, smc
 from cuthbert.filtering import filter
 from cuthbert.inference import (
     Filter,
