--- cuthbert/enkf/__init__.py
@@ -1,1 +1,0 @@
-from cuthbert.enkf import ensemble_kalman_filter, ensemble_rts_smoother
