+++ cuthbert/ensemble_kalman/__init__.py
@@ -1,0 +1,1 @@
+from cuthbert.ensemble_kalman import ensemble_kalman_filter, ensemble_rts_smoother
