--- cuthbertlib/enkf/__init__.py
@@ -1,2 +1,0 @@
-from cuthbertlib.enkf.filtering import predict, update
-from cuthbertlib.enkf.smoothing import update as smoother_update
