+++ cuthbertlib/ensemble_kalman/__init__.py
@@ -1,0 +1,3 @@
+from cuthbertlib.ensemble_kalman.filtering import predict
+from cuthbertlib.ensemble_kalman.filtering import update as filter_update
+from cuthbertlib.ensemble_kalman.smoothing import update as smoother_update
