--- docs/api_cuthbert/enkf.md
@@ -1,17 +1,0 @@
-# Ensemble Kalman Filter and Smoother
-
-- [Filtering interface](enkf/ensemble_kalman_filter.md)
-- [Smoothing interface](enkf/ensemble_rts_smoother.md)
-- [Callback types](enkf/types.md)
-
-::: cuthbert.enkf.ensemble_kalman_filter
-    options:
-      members:
-        - build_filter
-        - EnKFState
-
-::: cuthbert.enkf.ensemble_rts_smoother
-    options:
-      members:
-        - build_smoother
-        - EnRTSState
