--- docs/api_cuthbert/enkf/ensemble_kalman_filter.md
@@ -1,3 +1,0 @@
-# `ensemble_kalman_filter`
-
-::: cuthbert.enkf.ensemble_kalman_filter
