--- docs/api_cuthbert/enkf/ensemble_rts_smoother.md
@@ -1,3 +1,0 @@
-# `ensemble_rts_smoother`
-
-::: cuthbert.enkf.ensemble_rts_smoother
