--- docs/api_cuthbert/enkf/types.md
@@ -1,3 +1,0 @@
-# `types`
-
-::: cuthbert.enkf.types
