+++ docs/api_cuthbert/ensemble_kalman.md
@@ -1,0 +1,17 @@
+# Ensemble Kalman
+
+- [Filtering interface](ensemble_kalman/ensemble_kalman_filter.md)
+- [Smoothing interface](ensemble_kalman/ensemble_rts_smoother.md)
+- [Callback types](ensemble_kalman/types.md)
+
+::: cuthbert.ensemble_kalman.ensemble_kalman_filter
+    options:
+      members:
+        - build_filter
+        - EnKFState
+
+::: cuthbert.ensemble_kalman.ensemble_rts_smoother
+    options:
+      members:
+        - build_smoother
+        - EnRTSState
