+++ docs/api_cuthbert/ensemble_kalman/ensemble_kalman_filter.md
@@ -1,0 +1,3 @@
+# `ensemble_kalman_filter`
+
+::: cuthbert.ensemble_kalman.ensemble_kalman_filter
