+++ docs/api_cuthbert/ensemble_kalman/ensemble_rts_smoother.md
@@ -1,0 +1,3 @@
+# `ensemble_rts_smoother`
+
+::: cuthbert.ensemble_kalman.ensemble_rts_smoother
