+++ docs/api_cuthbert/ensemble_kalman/types.md
@@ -1,0 +1,3 @@
+# `types`
+
+::: cuthbert.ensemble_kalman.types
