--- docs/api_cuthbertlib/enkf.md
@@ -1,8 +1,0 @@
---8<-- "cuthbertlib/enkf/README.md"
-
-- [Filtering primitives](enkf/filtering.md)
-- [Smoothing primitives](enkf/smoothing.md)
-
-::: cuthbertlib.enkf.filtering
-
-::: cuthbertlib.enkf.smoothing
