--- docs/api_cuthbertlib/enkf/filtering.md
@@ -1,3 +1,0 @@
-# `filtering`
-
-::: cuthbertlib.enkf.filtering
