--- docs/api_cuthbertlib/enkf/smoothing.md
@@ -1,3 +1,0 @@
-# `smoothing`
-
-::: cuthbertlib.enkf.smoothing
