+++ docs/api_cuthbertlib/ensemble_kalman.md
@@ -1,0 +1,8 @@
+--8<-- "cuthbertlib/ensemble_kalman/README.md"
+
+- [Filtering primitives](ensemble_kalman/filtering.md)
+- [Smoothing primitives](ensemble_kalman/smoothing.md)
+
+::: cuthbertlib.ensemble_kalman.filtering
+
+::: cuthbertlib.ensemble_kalman.smoothing
