+++ docs/api_cuthbertlib/ensemble_kalman/filtering.md
@@ -1,0 +1,3 @@
+# `filtering`
+
+::: cuthbertlib.ensemble_kalman.filtering
