+++ docs/api_cuthbertlib/ensemble_kalman/smoothing.md
@@ -1,0 +1,3 @@
+# `smoothing`
+
+::: cuthbertlib.ensemble_kalman.smoothing
